% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_variavel_2.R
\name{df_variavel_2}
\alias{df_variavel_2}
\title{Data Frame com variável ao quadrado}
\usage{
df_variavel_2(df, variavel)
}
\arguments{
\item{df}{Um objeto do tipo `Data.Frame`}

\item{variavel}{O nome de uma coluna do `Data.Frame` especificado em `df`, que seja do tipo `numeric` ou `integer`}
}
\value{
Um Data Frame com mesmas colunas, porém com a coluna `variavel` elevada ao quadrado
}
\description{
Data Frame com variável ao quadrado
}
\examples{
df_variavel_2(iris, "Sepal.Length")
df_variavel_2(iris, "Petal.Width")

}
